safari-534.53-branch/LayoutTests/sputnik/Conformance/12_Statement/12.6_Iteration_Statements/12.6.2_The_while_statement/S12.6.2_A15-expected.txt

CONSOLE MESSAGE: line 78: SyntaxError: Parse error
S12.6.2_A15

PASS Expected parsing failure

TEST COMPLETE
